[AutoDiff] Move thunk generation utilities to Utils/Differentiation/Thunk.h.
