+++ include/swift/SILOptimizer/Utils/Differentiation/Common.h
@@ -1,4 +1,4 @@
-//===--- Common.h - Automatic Differentiation Common Utils ----*- C++ -*---===//
+//===--- Common.h - Automatic differentiation common utils ----*- C++ -*---===//
 //
 // This source file is part of the Swift.org open source project
 //
@@ -45,57 +45,6 @@ class DifferentiableActivityInfo;
 //===----------------------------------------------------------------------===//
 
 namespace autodiff {
-
-//===----------------------------------------------------------------------===//
-// Thunk helpers
-//===----------------------------------------------------------------------===//
-// These helpers are copied/adapted from SILGen. They should be refactored and
-// moved to a shared location.
-//===----------------------------------------------------------------------===//
-
-/// The thunk kinds used in the differentiation transform.
-enum class DifferentiationThunkKind {
-  /// A reabstraction thunk.
-  ///
-  /// Reabstraction thunks transform a function-typed value to another one with
-  /// different parameter/result abstraction patterns. This is identical to the
-  /// thunks generated by SILGen.
-  Reabstraction,
-
-  /// An index subset thunk.
-  ///
-  /// An index subset thunk is used transform JVP/VJPs into a version that is
-  /// "wrt" fewer differentiation parameters.
-  /// - Differentials of thunked JVPs use zero for non-requested differentiation
-  //    parameters.
-  /// - Pullbacks of thunked VJPs discard results for non-requested
-  ///   differentiation parameters.
-  IndexSubset
-};
-
-CanGenericSignature buildThunkSignature(SILFunction *fn, bool inheritGenericSig,
-                                        OpenedArchetypeType *openedExistential,
-                                        GenericEnvironment *&genericEnv,
-                                        SubstitutionMap &contextSubs,
-                                        SubstitutionMap &interfaceSubs,
-                                        ArchetypeType *&newArchetype);
-
-/// Build the type of a function transformation thunk.
-CanSILFunctionType buildThunkType(SILFunction *fn,
-                                  CanSILFunctionType &sourceType,
-                                  CanSILFunctionType &expectedType,
-                                  GenericEnvironment *&genericEnv,
-                                  SubstitutionMap &interfaceSubs,
-                                  bool withoutActuallyEscaping,
-                                  DifferentiationThunkKind thunkKind);
-
-/// Get or create a reabstraction thunk from `fromType` to `toType`, to be
-/// called in `caller`.
-SILFunction *getOrCreateReabstractionThunk(SILOptFunctionBuilder &fb,
-                                           SILModule &module, SILLocation loc,
-                                           SILFunction *caller,
-                                           CanSILFunctionType fromType,
-                                           CanSILFunctionType toType);
 
 /// Prints an "[AD] " prefix to `llvm::dbgs()` and returns the debug stream.
 /// This is being used to print short debug messages within the AD pass.
