+++ lib/SILOptimizer/Mandatory/Differentiation.cpp
@@ -53,6 +53,7 @@
 #include "swift/SILOptimizer/Utils/Differentiation/Common.h"
 #include "swift/SILOptimizer/Utils/Differentiation/JVPEmitter.h"
 #include "swift/SILOptimizer/Utils/Differentiation/LinearMapInfo.h"
+#include "swift/SILOptimizer/Utils/Differentiation/Thunk.h"
 #include "swift/SILOptimizer/Utils/Differentiation/VJPEmitter.h"
 #include "llvm/ADT/APSInt.h"
 #include "llvm/ADT/BreadthFirstIterator.h"
@@ -164,31 +165,6 @@ public:
   /// `SkipFoldingDifferentiableFunctionExtraction` flag for SIL testing
   /// purposes.
   void foldDifferentiableFunctionExtraction(DifferentiableFunctionInst *source);
-
-  /// Get or create a derivative function parameter index subset thunk from
-  /// `actualIndices` to `desiredIndices` for the given associated function
-  /// value and original function operand. Returns a pair of the parameter
-  /// index subset thunk and its interface substitution map (used to partially
-  /// apply the thunk).
-  /// Calls `getOrCreateSubsetParametersThunkForLinearMap` to thunk the linear
-  /// map returned by the derivative function.
-  std::pair<SILFunction *, SubstitutionMap>
-  getOrCreateSubsetParametersThunkForDerivativeFunction(
-      SILValue origFnOperand, SILValue derivativeFn,
-      AutoDiffDerivativeFunctionKind kind, SILAutoDiffIndices desiredIndices,
-      SILAutoDiffIndices actualIndices);
-
-  /// Get or create a derivative function parameter index subset thunk from
-  /// `actualIndices` to `desiredIndices` for the given associated function
-  /// value and original function operand. Returns a pair of the parameter
-  /// index subset thunk and its interface substitution map (used to partially
-  /// apply the thunk).
-  std::pair<SILFunction *, SubstitutionMap>
-  getOrCreateSubsetParametersThunkForLinearMap(
-      SILFunction *assocFn, CanSILFunctionType linearMapType,
-      CanSILFunctionType targetType, AutoDiffDerivativeFunctionKind kind,
-      SILAutoDiffIndices desiredIndices, SILAutoDiffIndices actualIndices);
-
 };
 
 } // end anonymous namespace
@@ -975,402 +951,6 @@ public:
 };
 } // end anonymous namespace
 
-std::pair<SILFunction *, SubstitutionMap>
-DifferentiationTransformer::getOrCreateSubsetParametersThunkForLinearMap(
-    SILFunction *parentThunk, CanSILFunctionType linearMapType,
-    CanSILFunctionType targetType, AutoDiffDerivativeFunctionKind kind,
-    SILAutoDiffIndices desiredIndices, SILAutoDiffIndices actualIndices) {
-  LLVM_DEBUG(getADDebugStream()
-             << "Getting a subset parameters thunk for " << linearMapType
-             << " from " << actualIndices << " to " << desiredIndices << '\n');
-
-  SubstitutionMap interfaceSubs;
-  GenericEnvironment *genericEnv = nullptr;
-  auto thunkType = buildThunkType(
-      parentThunk, linearMapType, targetType, genericEnv, interfaceSubs,
-      /*withoutActuallyEscaping*/ true,
-      DifferentiationThunkKind::Reabstraction);
-
-  // TODO(TF-685): Use more principled mangling for thunks.
-  std::string thunkName;
-  switch (kind) {
-    case AutoDiffDerivativeFunctionKind::JVP:
-      thunkName = "differential";
-      break;
-    case AutoDiffDerivativeFunctionKind::VJP:
-      thunkName = "pullback";
-  }
-  Mangle::ASTMangler mangler;
-  auto fromInterfaceType =
-      linearMapType->mapTypeOutOfContext()->getCanonicalType();
-  auto toInterfaceType = targetType->mapTypeOutOfContext()->getCanonicalType();
-  CanType dynamicSelfType;
-  thunkName = "AD__" + mangler.mangleReabstractionThunkHelper(
-      thunkType, fromInterfaceType, toInterfaceType, dynamicSelfType,
-      context.getModule().getSwiftModule()) + "_" + desiredIndices.mangle() + "_" +
-      thunkName;
-  thunkName += "_index_subset_thunk";
-
-  auto loc = parentThunk->getLocation();
-  SILOptFunctionBuilder fb(context.getTransform());
-  auto *thunk = fb.getOrCreateSharedFunction(
-      loc, thunkName, thunkType, IsBare, IsTransparent, IsSerialized,
-      ProfileCounter(), IsThunk, IsNotDynamic);
-
-  if (!thunk->empty())
-    return {thunk, interfaceSubs};
-
-  thunk->setGenericEnvironment(genericEnv);
-  thunk->setOwnershipEliminated();
-  auto *entry = thunk->createBasicBlock();
-  SILBuilder builder(entry);
-  createEntryArguments(thunk);
-
-  // Get arguments.
-  SmallVector<SILValue, 4> arguments;
-  SmallVector<AllocStackInst *, 4> localAllocations;
-
-  // Build a `.zero` argument for the given `Differentiable`-conforming type.
-  auto buildZeroArgument = [&](SILType zeroSILType) {
-    auto zeroSILObjType = zeroSILType.getObjectType();
-    auto zeroType = zeroSILType.getASTType();
-    auto *swiftMod = context.getModule().getSwiftModule();
-    auto tangentSpace = zeroType->getAutoDiffAssociatedTangentSpace(
-      LookUpConformanceInModule(swiftMod));
-    assert(tangentSpace && "No tangent space for this type");
-    switch (tangentSpace->getKind()) {
-    case VectorSpace::Kind::Vector: {
-      auto *buf = builder.createAllocStack(loc, zeroSILObjType);
-      localAllocations.push_back(buf);
-      emitZeroIntoBuffer(builder, zeroType, buf, loc);
-      if (zeroSILType.isAddress())
-        arguments.push_back(buf);
-      else {
-        auto *arg = builder.createLoad(loc, buf,
-                                       LoadOwnershipQualifier::Unqualified);
-        arguments.push_back(arg);
-      }
-      break;
-    }
-    case VectorSpace::Kind::Tuple: {
-      llvm_unreachable(
-          "Unimplemented: Handle zero initialization for tuples");
-    }
-    case VectorSpace::Kind::Function:
-      llvm_unreachable(
-          "Unimplemented: Emit thunks for abstracting zero initialization");
-    }
-  };
-
-  // `actualIndices` and `desiredIndices` are with respect to the original
-  // function. However, the differential parameters and pullback results may
-  // already be w.r.t. a subset. We create a map between the original function's
-  // actual parameter indices and the linear map's actual indices.
-  // Example:
-  //   Original: (T0, T1, T2) -> R
-  //   Actual indices: 0, 2
-  //   Original differential: (T0, T2) -> R
-  //   Original pullback: R -> (T0, T2)
-  //   Desired indices w.r.t. original: 2
-  //   Desired indices w.r.t. linear map: 1
-  SmallVector<unsigned, 4> actualParamIndicesMap(
-      actualIndices.parameters->getCapacity(), UINT_MAX);
-  {
-    unsigned indexInBitVec = 0;
-    for (auto index : actualIndices.parameters->getIndices()) {
-      actualParamIndicesMap[index] = indexInBitVec;
-      indexInBitVec++;
-    }
-  }
-  auto mapOriginalParameterIndex = [&](unsigned index) -> unsigned {
-    auto mappedIndex = actualParamIndicesMap[index];
-    assert(mappedIndex < actualIndices.parameters->getCapacity());
-    return mappedIndex;
-  };
-
-  switch (kind) {
-  // Differential arguments are:
-  // - All indirect results, followed by:
-  // - An interleaving of:
-  //   - Thunk arguments (when parameter index is in both desired and actual
-  //     indices).
-  //   - Zeros (when parameter is not in desired indices).
-  case AutoDiffDerivativeFunctionKind::JVP: {
-    // Forward all indirect results.
-    arguments.append(thunk->getIndirectResults().begin(),
-                     thunk->getIndirectResults().end());
-    auto toArgIter = thunk->getArgumentsWithoutIndirectResults().begin();
-    auto useNextArgument = [&]() {
-      arguments.push_back(*toArgIter++);
-    };
-    // Iterate over actual indices.
-    for (unsigned i : actualIndices.parameters->getIndices()) {
-      // If index is desired, use next argument.
-      if (desiredIndices.isWrtParameter(i)) {
-        useNextArgument();
-      }
-      // Otherwise, construct and use a zero argument.
-      else {
-        auto zeroSILType =
-            linearMapType->getParameters()[mapOriginalParameterIndex(i)]
-                .getSILStorageInterfaceType();
-        buildZeroArgument(zeroSILType);
-      }
-    }
-    break;
-  }
-  // Pullback arguments are:
-  // - An interleaving of:
-  //   - Thunk indirect results (when parameter index is in both desired and
-  //     actual indices).
-  //   - Zeros (when parameter is not in desired indices).
-  // - All actual arguments.
-  case AutoDiffDerivativeFunctionKind::VJP: {
-    auto toIndirectResultsIter = thunk->getIndirectResults().begin();
-    auto useNextResult = [&]() {
-      arguments.push_back(*toIndirectResultsIter++);
-    };
-    // Iterate over actual indices.
-    for (unsigned i : actualIndices.parameters->getIndices()) {
-      auto resultInfo =
-          linearMapType->getResults()[mapOriginalParameterIndex(i)];
-      // Skip direct results. Only indirect results are relevant as arguments.
-      if (resultInfo.isFormalDirect())
-        continue;
-      // If index is desired, use next indirect result.
-      if (desiredIndices.isWrtParameter(i)) {
-        useNextResult();
-        continue;
-      }
-      // Otherwise, construct and use an uninitialized indirect result.
-      auto *indirectResult =
-          builder.createAllocStack(loc, resultInfo.getSILStorageInterfaceType());
-      localAllocations.push_back(indirectResult);
-      arguments.push_back(indirectResult);
-    }
-    // Foward all actual non-indirect-result arguments.
-    arguments.append(thunk->getArgumentsWithoutIndirectResults().begin(),
-                     thunk->getArgumentsWithoutIndirectResults().end() - 1);
-    break;
-  }
-  }
-
-  // Get the linear map thunk argument and apply it.
-  auto *linearMap = thunk->getArguments().back();
-  auto *ai = builder.createApply(
-      loc, linearMap, SubstitutionMap(), arguments, /*isNonThrowing*/ false);
-
-  // If differential thunk, deallocate local allocations and directly return
-  // `apply` result.
-  if (kind == AutoDiffDerivativeFunctionKind::JVP) {
-    for (auto *alloc : llvm::reverse(localAllocations))
-      builder.createDeallocStack(loc, alloc);
-    builder.createReturn(loc, ai);
-    return {thunk, interfaceSubs};
-  }
-
-  // If pullback thunk, return only the desired results and clean up the
-  // undesired results.
-  SmallVector<SILValue, 8> pullbackDirectResults;
-  extractAllElements(ai, builder, pullbackDirectResults);
-  SmallVector<SILValue, 8> allResults;
-  collectAllActualResultsInTypeOrder(ai, pullbackDirectResults, allResults);
-
-  SmallVector<SILValue, 8> results;
-  for (unsigned i : actualIndices.parameters->getIndices()) {
-    // If result is desired:
-    // - Do nothing if result is indirect.
-    //   (It was already forwarded to the `apply` instruction).
-    // - Push it to `results` if result is direct.
-    auto result = allResults[mapOriginalParameterIndex(i)];
-    if (desiredIndices.isWrtParameter(i)) {
-      if (result->getType().isObject())
-        results.push_back(result);
-    }
-    // Otherwise, cleanup the unused results.
-    else {
-      if (result->getType().isAddress())
-        builder.emitDestroyAddrAndFold(loc, result);
-      else
-        builder.emitDestroyValueOperation(loc, result);
-    }
-  }
-  // Deallocate local allocations and return final direct result.
-  for (auto *alloc : llvm::reverse(localAllocations))
-    builder.createDeallocStack(loc, alloc);
-  auto result = joinElements(results, builder, loc);
-  builder.createReturn(loc, result);
-
-  context.recordGeneratedFunction(thunk);
-  return {thunk, interfaceSubs};
-}
-
-std::pair<SILFunction *, SubstitutionMap>
-DifferentiationTransformer::getOrCreateSubsetParametersThunkForDerivativeFunction(
-    SILValue origFnOperand, SILValue derivativeFn,
-    AutoDiffDerivativeFunctionKind kind, SILAutoDiffIndices desiredIndices,
-    SILAutoDiffIndices actualIndices) {
-  LLVM_DEBUG(getADDebugStream()
-             << "Getting a subset parameters thunk for derivative function "
-             << derivativeFn << " of the original function " << origFnOperand
-             << " from " << actualIndices << " to " << desiredIndices << '\n');
-
-  auto origFnType = origFnOperand->getType().castTo<SILFunctionType>();
-  auto &module = context.getModule();
-  auto lookupConformance = LookUpConformanceInModule(module.getSwiftModule());
-
-  // Compute target type for thunking.
-  auto derivativeFnType = derivativeFn->getType().castTo<SILFunctionType>();
-  auto targetType = origFnType->getAutoDiffDerivativeFunctionType(
-      desiredIndices.parameters, desiredIndices.source, kind, module.Types,
-      lookupConformance);
-  auto *caller = derivativeFn->getFunction();
-  if (targetType->hasArchetype()) {
-    auto substTargetType = caller->mapTypeIntoContext(
-        targetType->mapTypeOutOfContext())->getCanonicalType();
-    targetType = SILType::getPrimitiveObjectType(substTargetType)
-        .castTo<SILFunctionType>();
-  }
-  assert(derivativeFnType->getNumParameters() == targetType->getNumParameters());
-  assert(derivativeFnType->getNumResults() == targetType->getNumResults());
-
-  // Build thunk type.
-  SubstitutionMap interfaceSubs;
-  GenericEnvironment *genericEnv = nullptr;
-  auto thunkType = buildThunkType(
-      derivativeFn->getFunction(), derivativeFnType, targetType, genericEnv,
-      interfaceSubs, /*withoutActuallyEscaping*/ false,
-      DifferentiationThunkKind::IndexSubset);
-
-  // FIXME: The logic for resolving `assocRef` does not reapply function
-  // conversions, which is problematic if `derivativeFn` is a `partial_apply`
-  // instruction.
-  StringRef origName;
-  if (auto *origFnRef =
-          peerThroughFunctionConversions<FunctionRefInst>(origFnOperand)) {
-    origName = origFnRef->getInitiallyReferencedFunction()->getName();
-  } else if (auto *origMethodInst =
-                 peerThroughFunctionConversions<MethodInst>(origFnOperand)) {
-    origName = origMethodInst->getMember().getAnyFunctionRef()
-        ->getAbstractFunctionDecl()->getNameStr();
-  }
-  assert(!origName.empty() && "Original function name could not be resolved");
-  // TODO(TF-685): Use more principled mangling for thunks.
-  std::string thunkName;
-  switch (kind) {
-    case AutoDiffDerivativeFunctionKind::JVP:
-      thunkName = "jvp";
-      break;
-    case AutoDiffDerivativeFunctionKind::VJP:
-      thunkName = "vjp";
-  }
-  Mangle::ASTMangler mangler;
-  auto fromInterfaceType =
-      derivativeFnType->mapTypeOutOfContext()->getCanonicalType();
-  auto toInterfaceType = targetType->mapTypeOutOfContext()->getCanonicalType();
-  CanType dynamicSelfType;
-  thunkName = "AD__orig_" + origName.str() + "_" +
-      mangler.mangleReabstractionThunkHelper(
-          thunkType, fromInterfaceType, toInterfaceType, dynamicSelfType,
-          module.getSwiftModule()) + "_" + desiredIndices.mangle() + "_" +
-          thunkName;
-  thunkName += "_subset_parameters_thunk";
-
-  auto loc = origFnOperand.getLoc();
-  SILOptFunctionBuilder fb(context.getTransform());
-  auto *thunk = fb.getOrCreateSharedFunction(
-      loc, thunkName, thunkType, IsBare, IsTransparent, caller->isSerialized(),
-      ProfileCounter(), IsThunk, IsNotDynamic);
-
-  if (!thunk->empty())
-    return {thunk, interfaceSubs};
-
-  thunk->setOwnershipEliminated();
-  thunk->setGenericEnvironment(genericEnv);
-  auto *entry = thunk->createBasicBlock();
-  SILBuilder builder(entry);
-  createEntryArguments(thunk);
-
-  SubstitutionMap assocSubstMap;
-  if (auto *partialApply = dyn_cast<PartialApplyInst>(derivativeFn))
-    assocSubstMap = partialApply->getSubstitutionMap();
-
-  // FIXME: The logic for resolving `assocRef` does not reapply function
-  // conversions, which is problematic if `derivativeFn` is a `partial_apply`
-  // instruction.
-  SILValue assocRef;
-  if (auto *derivativeFnRef =
-          peerThroughFunctionConversions<FunctionRefInst>(derivativeFn)) {
-    auto *assoc = derivativeFnRef->getReferencedFunctionOrNull();
-    assocRef = builder.createFunctionRef(loc, assoc);
-  } else if (auto *assocMethodInst =
-                 peerThroughFunctionConversions<WitnessMethodInst>(derivativeFn)) {
-    assocRef = builder.createWitnessMethod(
-        loc, assocMethodInst->getLookupType(),
-        assocMethodInst->getConformance(), assocMethodInst->getMember(),
-        thunk->mapTypeIntoContext(assocMethodInst->getType()));
-  } else if (auto *assocMethodInst =
-                 peerThroughFunctionConversions<ClassMethodInst>(derivativeFn)) {
-    auto classOperand = thunk->getArgumentsWithoutIndirectResults().back();
-    auto classOperandType = assocMethodInst->getOperand()->getType();
-    assert(classOperand->getType() == classOperandType);
-    assocRef = builder.createClassMethod(
-        loc, classOperand, assocMethodInst->getMember(),
-        thunk->mapTypeIntoContext(assocMethodInst->getType()));
-  } else if (auto *diffWitFn = peerThroughFunctionConversions<
-                 DifferentiabilityWitnessFunctionInst>(derivativeFn)) {
-    assocRef = builder.createDifferentiabilityWitnessFunction(
-        loc, diffWitFn->getWitnessKind(), diffWitFn->getWitness());
-  }
-  assert(assocRef && "Expected derivative function to be resolved");
-
-  assocSubstMap = assocSubstMap.subst(thunk->getForwardingSubstitutionMap());
-  derivativeFnType = assocRef->getType().castTo<SILFunctionType>();
-
-  SmallVector<SILValue, 4> arguments;
-  arguments.append(thunk->getArguments().begin(), thunk->getArguments().end());
-  assert(arguments.size() == derivativeFnType->getNumParameters() +
-                                 derivativeFnType->getNumIndirectFormalResults());
-  auto *apply = builder.createApply(
-      loc, assocRef, assocSubstMap, arguments, /*isNonThrowing*/ false);
-
-  // Extract all direct results.
-  SmallVector<SILValue, 8> directResults;
-  extractAllElements(apply, builder, directResults);
-  auto originalDirectResults = ArrayRef<SILValue>(directResults).drop_back(1);
-  auto originalDirectResult =
-      joinElements(originalDirectResults, builder, apply->getLoc());
-  auto linearMap = directResults.back();
-
-  auto linearMapType = linearMap->getType().castTo<SILFunctionType>();
-  auto linearMapTargetType = targetType->getResults().back().getSILStorageInterfaceType()
-      .castTo<SILFunctionType>();
-
-  SILFunction *linearMapThunk;
-  SubstitutionMap linearMapSubs;
-  std::tie(linearMapThunk, linearMapSubs) =
-      getOrCreateSubsetParametersThunkForLinearMap(
-          thunk, linearMapType, linearMapTargetType, kind,
-          desiredIndices, actualIndices);
-
-  auto *linearMapThunkFRI = builder.createFunctionRef(loc, linearMapThunk);
-  auto *thunkedLinearMap = builder.createPartialApply(
-      loc, linearMapThunkFRI, linearMapSubs, {linearMap},
-      ParameterConvention::Direct_Guaranteed);
-
-  assert(origFnType->getResults().size() == 1);
-  if (origFnType->getResults().front().isFormalDirect()) {
-    auto result = joinElements(
-        {originalDirectResult, thunkedLinearMap}, builder, loc);
-    builder.createReturn(loc, result);
-  } else {
-    builder.createReturn(loc, thunkedLinearMap);
-  }
-
-  context.recordGeneratedFunction(thunk);
-  return {thunk, interfaceSubs};
-}
-
 SILValue DifferentiationTransformer::promoteToDifferentiableFunction(
     DifferentiableFunctionInst *dfi, SILBuilder &builder, SILLocation loc,
     DifferentiationInvoker invoker) {
@@ -1470,7 +1050,7 @@ SILValue DifferentiationTransformer::promoteToDifferentiableFunction(
   SmallVector<SILValue, 2> derivativeFns;
   SmallVector<AllocStackInst *, 2> newBuffersToDealloc;
   for (auto derivativeFnKind : {AutoDiffDerivativeFunctionKind::JVP,
-                           AutoDiffDerivativeFunctionKind::VJP}) {
+                                AutoDiffDerivativeFunctionKind::VJP}) {
     auto derivativeFnAndIndices = emitDerivativeFunctionReference(
         *this, builder, desiredIndices, derivativeFnKind, origFnOperand, invoker,
         newBuffersToDealloc);
@@ -1523,9 +1103,10 @@ SILValue DifferentiationTransformer::promoteToDifferentiableFunction(
       assert(actualIndices.parameters->isSupersetOf(extendedDesiredIndices));
       SILFunction *thunk;
       SubstitutionMap interfaceSubs;
+      SILOptFunctionBuilder fb(transform);
       std::tie(thunk, interfaceSubs) =
           getOrCreateSubsetParametersThunkForDerivativeFunction(
-              origFnOperand, derivativeFn, derivativeFnKind, desiredIndices,
+              fb, origFnOperand, derivativeFn, derivativeFnKind, desiredIndices,
               actualIndices);
       auto *thunkFRI = builder.createFunctionRef(loc, thunk);
       if (auto genSig =
