+++ lib/SILOptimizer/Utils/Differentiation/Common.cpp
@@ -1,4 +1,4 @@
-//===--- Differentiation.cpp - SIL Automatic Differentiation --*- C++ -*---===//
+//===--- Common.cpp - Automatic differentiation common utils --*- C++ -*---===//
 //
 // This source file is part of the Swift.org open source project
 //
@@ -224,401 +224,6 @@ void emitZeroIntoBuffer(SILBuilder &builder, CanType type,
   builder.createApply(loc, getter, subMap, {bufferAccess, metatype},
                       /*isNonThrowing*/ false);
   builder.emitDestroyValueOperation(loc, getter);
-}
-
-//===----------------------------------------------------------------------===//
-// Thunk helpers
-//===----------------------------------------------------------------------===//
-// These helpers are copied/adapted from SILGen. They should be refactored and
-// moved to a shared location.
-//===----------------------------------------------------------------------===//
-
-CanGenericSignature buildThunkSignature(SILFunction *fn, bool inheritGenericSig,
-                                        OpenedArchetypeType *openedExistential,
-                                        GenericEnvironment *&genericEnv,
-                                        SubstitutionMap &contextSubs,
-                                        SubstitutionMap &interfaceSubs,
-                                        ArchetypeType *&newArchetype) {
-  // If there's no opened existential, we just inherit the generic environment
-  // from the parent function.
-  if (openedExistential == nullptr) {
-    auto genericSig = fn->getLoweredFunctionType()->getSubstGenericSignature();
-    genericEnv = fn->getGenericEnvironment();
-    interfaceSubs = fn->getForwardingSubstitutionMap();
-    contextSubs = interfaceSubs;
-    return genericSig;
-  }
-
-  auto &ctx = fn->getASTContext();
-  GenericSignatureBuilder builder(ctx);
-
-  // Add the existing generic signature.
-  int depth = 0;
-  if (inheritGenericSig) {
-    if (auto genericSig =
-            fn->getLoweredFunctionType()->getSubstGenericSignature()) {
-      builder.addGenericSignature(genericSig);
-      depth = genericSig->getGenericParams().back()->getDepth() + 1;
-    }
-  }
-
-  // Add a new generic parameter to replace the opened existential.
-  auto *newGenericParam = GenericTypeParamType::get(depth, 0, ctx);
-
-  builder.addGenericParameter(newGenericParam);
-  Requirement newRequirement(RequirementKind::Conformance, newGenericParam,
-                             openedExistential->getOpenedExistentialType());
-  auto source =
-      GenericSignatureBuilder::FloatingRequirementSource::forAbstract();
-  builder.addRequirement(newRequirement, source, nullptr);
-
-  auto genericSig = std::move(builder).computeGenericSignature(
-      SourceLoc(), /*allowConcreteGenericParams=*/true);
-  genericEnv = genericSig->getGenericEnvironment();
-
-  newArchetype = genericEnv->mapTypeIntoContext(newGenericParam)
-      ->castTo<ArchetypeType>();
-
-  // Calculate substitutions to map the caller's archetypes to the thunk's
-  // archetypes.
-  if (auto calleeGenericSig =
-          fn->getLoweredFunctionType()->getSubstGenericSignature()) {
-    contextSubs = SubstitutionMap::get(
-        calleeGenericSig,
-        [&](SubstitutableType *type) -> Type {
-          return genericEnv->mapTypeIntoContext(type);
-        },
-        MakeAbstractConformanceForGenericType());
-  }
-
-  // Calculate substitutions to map interface types to the caller's archetypes.
-  interfaceSubs = SubstitutionMap::get(
-      genericSig,
-      [&](SubstitutableType *type) -> Type {
-        if (type->isEqual(newGenericParam))
-          return openedExistential;
-        return fn->mapTypeIntoContext(type);
-      },
-      MakeAbstractConformanceForGenericType());
-
-  return genericSig->getCanonicalSignature();
-}
-
-CanSILFunctionType buildThunkType(SILFunction *fn,
-                                  CanSILFunctionType &sourceType,
-                                  CanSILFunctionType &expectedType,
-                                  GenericEnvironment *&genericEnv,
-                                  SubstitutionMap &interfaceSubs,
-                                  bool withoutActuallyEscaping,
-                                  DifferentiationThunkKind thunkKind) {
-  assert(!expectedType->isPolymorphic());
-  assert(!sourceType->isPolymorphic());
-
-  auto &module = fn->getModule();
-  auto origType = sourceType;
-
-  // Cannot build a reabstraction thunk without context. Ownership semantics
-  // on the result type are required.
-  if (thunkKind == DifferentiationThunkKind::Reabstraction)
-    assert(expectedType->getExtInfo().hasContext());
-
-  // This may inherit @noescape from the expected type. The `@noescape`
-  // attribute is only stripped when using this type to materialize a new decl.
-  // Use `@convention(thin)` if:
-  // - Building a reabstraction thunk type.
-  // - Building an index subset thunk type, where the expected type has context
-  //   (i.e. is `@convention(thick)`).
-  auto extInfo = expectedType->getExtInfo();
-  if (thunkKind == DifferentiationThunkKind::Reabstraction ||
-      extInfo.hasContext()) {
-    extInfo = extInfo.withRepresentation(
-        SILFunctionType::Representation::Thin);
-  }
-  if (withoutActuallyEscaping)
-    extInfo = extInfo.withNoEscape(false);
-
-  // Does the thunk type involve archetypes other than opened existentials?
-  bool hasArchetypes = false;
-  // Does the thunk type involve an open existential type?
-  CanOpenedArchetypeType openedExistential;
-  auto archetypeVisitor = [&](CanType t) {
-    if (auto archetypeTy = dyn_cast<OpenedArchetypeType>(t)) {
-      if (auto opened = dyn_cast<OpenedArchetypeType>(archetypeTy)) {
-        assert((openedExistential == CanArchetypeType() ||
-                openedExistential == opened) &&
-               "one too many open existentials");
-        openedExistential = opened;
-      } else {
-        hasArchetypes = true;
-      }
-    }
-  };
-
-  // Use the generic signature from the context if the thunk involves
-  // generic parameters.
-  CanGenericSignature genericSig;
-  SubstitutionMap contextSubs;
-  ArchetypeType *newArchetype = nullptr;
-
-  if (expectedType->hasArchetype() || sourceType->hasArchetype()) {
-    expectedType.visit(archetypeVisitor);
-    sourceType.visit(archetypeVisitor);
-    genericSig = buildThunkSignature(
-        fn, hasArchetypes, openedExistential, genericEnv, contextSubs,
-        interfaceSubs, newArchetype);
-  }
-
-  // Utility function to apply contextSubs, and also replace the
-  // opened existential with the new archetype.
-  auto substIntoThunkContext = [&](CanType t) -> CanType {
-    return t.subst(
-        [&](SubstitutableType *type) -> Type {
-          if (CanType(type) == openedExistential)
-            return newArchetype;
-          return Type(type).subst(contextSubs);
-        },
-        LookUpConformanceInSubstitutionMap(contextSubs),
-        SubstFlags::AllowLoweredTypes)->getCanonicalType();
-  };
-
-  sourceType = cast<SILFunctionType>(substIntoThunkContext(sourceType));
-  expectedType = cast<SILFunctionType>(substIntoThunkContext(expectedType));
-
-  // If our parent function was pseudogeneric, this thunk must also be
-  // pseudogeneric, since we have no way to pass generic parameters.
-  if (genericSig)
-    if (origType->isPseudogeneric())
-      extInfo = extInfo.withIsPseudogeneric();
-
-  // Add the function type as the parameter.
-  auto contextConvention =
-      SILType::getPrimitiveObjectType(sourceType).isTrivial(*fn)
-          ? ParameterConvention::Direct_Unowned
-          : ParameterConvention::Direct_Guaranteed;
-  SmallVector<SILParameterInfo, 4> params;
-  params.append(expectedType->getParameters().begin(),
-                expectedType->getParameters().end());
-  // Add reabstraction function parameter only if building a reabstraction thunk
-  // type.
-  if (thunkKind == DifferentiationThunkKind::Reabstraction)
-    params.push_back({sourceType, sourceType->getExtInfo().hasContext()
-                                      ? contextConvention
-                                      : ParameterConvention::Direct_Unowned});
-
-  // Map the parameter and expected types out of context to get the interface
-  // type of the thunk.
-  SmallVector<SILParameterInfo, 4> interfaceParams;
-  interfaceParams.reserve(params.size());
-  for (auto &param : params) {
-    auto paramIfaceTy = param.getInterfaceType()->mapTypeOutOfContext();
-    interfaceParams.push_back(SILParameterInfo(
-        paramIfaceTy->getCanonicalType(genericSig), param.getConvention()));
-  }
-
-  SmallVector<SILYieldInfo, 4> interfaceYields;
-  for (auto &yield : expectedType->getYields()) {
-    auto yieldIfaceTy = yield.getInterfaceType()->mapTypeOutOfContext();
-    auto interfaceYield =
-        yield.getWithInterfaceType(yieldIfaceTy->getCanonicalType(genericSig));
-    interfaceYields.push_back(interfaceYield);
-  }
-
-  SmallVector<SILResultInfo, 4> interfaceResults;
-  for (auto &result : expectedType->getResults()) {
-    auto resultIfaceTy = result.getInterfaceType()->mapTypeOutOfContext();
-    auto interfaceResult =
-        result.getWithInterfaceType(resultIfaceTy->getCanonicalType(genericSig));
-    interfaceResults.push_back(interfaceResult);
-  }
-
-  Optional<SILResultInfo> interfaceErrorResult;
-  if (expectedType->hasErrorResult()) {
-    auto errorResult = expectedType->getErrorResult();
-    auto errorIfaceTy = errorResult.getInterfaceType()->mapTypeOutOfContext();
-    interfaceErrorResult =
-        SILResultInfo(errorIfaceTy->getCanonicalType(genericSig),
-                      expectedType->getErrorResult().getConvention());
-  }
-
-  // The type of the thunk function.
-  return SILFunctionType::get(
-      genericSig, extInfo, expectedType->getCoroutineKind(),
-      ParameterConvention::Direct_Unowned, interfaceParams, interfaceYields,
-      interfaceResults, interfaceErrorResult, {}, false, module.getASTContext());
-}
-
-SILFunction *getOrCreateReabstractionThunk(SILOptFunctionBuilder &fb,
-                                           SILModule &module, SILLocation loc,
-                                           SILFunction *caller,
-                                           CanSILFunctionType fromType,
-                                           CanSILFunctionType toType) {
-  SubstitutionMap interfaceSubs;
-  GenericEnvironment *genericEnv = nullptr;
-  auto thunkType = buildThunkType(
-      caller, fromType, toType, genericEnv, interfaceSubs,
-      /*withoutActuallyEscaping*/ false,
-      DifferentiationThunkKind::Reabstraction);
-  auto thunkDeclType =
-      thunkType->getWithExtInfo(thunkType->getExtInfo().withNoEscape(false));
-
-  auto fromInterfaceType = fromType->mapTypeOutOfContext()->getCanonicalType();
-  auto toInterfaceType = toType->mapTypeOutOfContext()->getCanonicalType();
-
-  Mangle::ASTMangler mangler;
-  std::string name = mangler.mangleReabstractionThunkHelper(
-      thunkType, fromInterfaceType, toInterfaceType,
-      Type(), module.getSwiftModule());
-
-  // FIXME(TF-989): Mark reabstraction thunks as transparent. This requires
-  // generating ossa reabstraction thunks so that they can be inlined during
-  // mandatory inlining when `-enable-strip-ownership-after-serialization` is
-  // true and ownership model eliminator is not run after differentiation.
-  auto *thunk = fb.getOrCreateSharedFunction(
-      loc, name, thunkDeclType, IsBare, IsNotTransparent, IsSerialized,
-      ProfileCounter(), IsReabstractionThunk, IsNotDynamic);
-  if (!thunk->empty())
-    return thunk;
-
-  thunk->setGenericEnvironment(genericEnv);
-  thunk->setOwnershipEliminated();
-  auto *entry = thunk->createBasicBlock();
-  SILBuilder builder(entry);
-  createEntryArguments(thunk);
-
-  SILFunctionConventions fromConv(fromType, module);
-  SILFunctionConventions toConv(toType, module);
-  assert(toConv.useLoweredAddresses());
-
-  auto *fnArg = thunk->getArgumentsWithoutIndirectResults().back();
-
-  SmallVector<SILValue, 4> arguments;
-  auto toArgIter = thunk->getArguments().begin();
-  auto useNextArgument = [&]() {
-    arguments.push_back(*toArgIter++);
-  };
-
-  SmallVector<AllocStackInst *, 4> localAllocations;
-  auto createAllocStack = [&](SILType type) {
-    auto *alloc = builder.createAllocStack(loc, type);
-    localAllocations.push_back(alloc);
-    return alloc;
-  };
-
-  // Handle indirect results.
-  assert(fromType->getNumResults() == toType->getNumResults());
-  for (unsigned resIdx : range(toType->getNumResults())) {
-    auto fromRes = fromConv.getResults()[resIdx];
-    auto toRes = toConv.getResults()[resIdx];
-    // No abstraction mismatch.
-    if (fromRes.isFormalIndirect() == toRes.isFormalIndirect()) {
-      // If result types are indirect, directly pass as next argument.
-      if (toRes.isFormalIndirect())
-        useNextArgument();
-      continue;
-    }
-    // Convert indirect result to direct result.
-    if (fromRes.isFormalIndirect()) {
-      SILType resultTy = fromConv.getSILType(fromRes);
-      assert(resultTy.isAddress());
-      auto *indRes = createAllocStack(resultTy);
-      arguments.push_back(indRes);
-      continue;
-    }
-    // Convert direct result to indirect result.
-    // Increment thunk argument iterator; reabstraction handled later.
-    toArgIter++;
-  }
-
-  // Reabstract parameters.
-  assert(toType->getNumParameters() == fromType->getNumParameters());
-  for (unsigned paramIdx : range(toType->getNumParameters())) {
-    auto fromParam = fromConv.getParameters()[paramIdx];
-    auto toParam = toConv.getParameters()[paramIdx];
-    // No abstraction mismatch. Directly use next argument.
-    if (fromParam.isFormalIndirect() == toParam.isFormalIndirect()) {
-      useNextArgument();
-      continue;
-    }
-    // Convert indirect parameter to direct parameter.
-    if (fromParam.isFormalIndirect()) {
-      auto paramTy = fromConv.getSILType(fromType->getParameters()[paramIdx]);
-      if (!paramTy.hasArchetype())
-        paramTy = thunk->mapTypeIntoContext(paramTy);
-      assert(paramTy.isAddress());
-      auto *toArg = *toArgIter++;
-      auto *buf = createAllocStack(toArg->getType());
-      builder.createStore(loc, toArg, buf,
-                          StoreOwnershipQualifier::Unqualified);
-      arguments.push_back(buf);
-      continue;
-    }
-    // Convert direct parameter to indirect parameter.
-    assert(toParam.isFormalIndirect());
-    auto *toArg = *toArgIter++;
-    auto *load = builder.createLoad(loc, toArg,
-                                    LoadOwnershipQualifier::Unqualified);
-    arguments.push_back(load);
-  }
-
-  auto *apply = builder.createApply(
-      loc, fnArg, SubstitutionMap(), arguments, /*isNonThrowing*/ false);
-
-  // Get return elements.
-  SmallVector<SILValue, 4> results;
-  // Extract all direct results.
-  SmallVector<SILValue, 4> directResults;
-  extractAllElements(apply, builder, directResults);
-
-  auto fromDirResultsIter = directResults.begin();
-  auto fromIndResultsIter = apply->getIndirectSILResults().begin();
-  auto toIndResultsIter = thunk->getIndirectResults().begin();
-  // Reabstract results.
-  for (unsigned resIdx : range(toType->getNumResults())) {
-    auto fromRes = fromConv.getResults()[resIdx];
-    auto toRes = toConv.getResults()[resIdx];
-    // No abstraction mismatch.
-    if (fromRes.isFormalIndirect() == toRes.isFormalIndirect()) {
-      // If result types are direct, add call result as direct thunk result.
-      if (toRes.isFormalDirect())
-        results.push_back(*fromDirResultsIter++);
-      // If result types are indirect, increment indirect result iterators.
-      else {
-        ++fromIndResultsIter;
-        ++toIndResultsIter;
-      }
-      continue;
-    }
-    // Load direct results from indirect results.
-    if (fromRes.isFormalIndirect()) {
-      auto indRes = *fromIndResultsIter++;
-      auto *load = builder.createLoad(loc, indRes,
-                                      LoadOwnershipQualifier::Unqualified);
-      results.push_back(load);
-      continue;
-    }
-    // Store direct results to indirect results.
-    assert(toRes.isFormalIndirect());
-    SILType resultTy = toConv.getSILType(toRes);
-    assert(resultTy.isAddress());
-    auto indRes = *toIndResultsIter++;
-    builder.createStore(loc, *fromDirResultsIter++, indRes,
-                        StoreOwnershipQualifier::Unqualified);
-  }
-  auto retVal = joinElements(results, builder, loc);
-
-  // Deallocate local allocations.
-  for (auto *alloc : llvm::reverse(localAllocations))
-    builder.createDeallocStack(loc, alloc);
-
-  // Create return.
-  builder.createReturn(loc, retVal);
-
-  LLVM_DEBUG(auto &s = getADDebugStream() << "Created reabstraction thunk.\n";
-             s << "  From type: " << fromType << '\n';
-             s << "  To type: " << toType << '\n';
-             s << '\n' << *thunk);
-
-  return thunk;
 }
 
 //===----------------------------------------------------------------------===//
