+++ lib/SILOptimizer/Utils/Differentiation/VJPEmitter.cpp
@@ -24,6 +24,7 @@
 #include "swift/SILOptimizer/Utils/Differentiation/ADContext.h"
 #include "swift/SILOptimizer/Utils/Differentiation/PullbackEmitter.h"
 #include "swift/SILOptimizer/Utils/Differentiation/VJPEmitter.h"
+#include "swift/SILOptimizer/Utils/Differentiation/Thunk.h"
 
 namespace swift {
 namespace autodiff {
